// app-original/src/app/interfaces/dataIndivideo.ts
export interface DataIndivideo {
    edad: number;
    sexo: string;
    ingresoBrutoMensual: number;
    saldoActualEnAhorro: number;
    densidad: number;
    pensionEstimada: number;
    pensionEstimada12Meses: number;
    pensionEstimada12MesesAnosExtra: number;
    mejorarPension: boolean;
    metaPension: number;
    pensionMaximaSolidaria: number;
    idCliente12Meses: number;
    idCliente12MesesAnosExtra: number;
    ahorroMaximaAPV: number;
    programmedRetirement: boolean;
    retirementFactor: number;
    email: string;
    valorApv: number;
    pensionEstimada12MesesConApv: number;
    pensionEstimada12MesesAnosExtraConApv: number;
}